lots of fixes, write manage.py command to update user locations following new algorithm
